refactor(modules): extract token-accounting bounded context
The 7th BC characterized during the Event Storming session of 2026-05-19.
Moves Claude token consumption and cost tracking into src/modules/token-accounting/.

- src/entities/tokenUsage/            → modules/token-accounting/entities/tokenUsage/
- src/usecases/trackTokenUsage/       → modules/token-accounting/usecases/trackTokenUsage/
- src/usecases/summarizeTokenUsage/   → modules/token-accounting/usecases/summarizeTokenUsage/
- src/interface-adapters/gateways/tokenUsage/ → modules/token-accounting/interface-adapters/gateways/tokenUsage/
