+++ src/tests/factories/tokenUsage.factory.ts
@@ -1,4 +1,4 @@
-import type { TokenUsage, TokenUsageRecord } from '@/entities/tokenUsage/tokenUsage.schema.js';
+import type { TokenUsage, TokenUsageRecord } from '@/modules/token-accounting/entities/tokenUsage/tokenUsage.schema.js';
 
 export class TokenUsageFactory {
   static create(overrides?: Partial<TokenUsage>): TokenUsage {
